updated namespace search and add new index.ts
